[Gold V] Title: DKSH를 찾아라, Time: 0 ms, Memory: 2020 KB, Score: 100 point -BaekjoonHub

+++ 백준/Gold/25427. DKSH를 찾아라/DKSH를 찾아라.cc
@@ -66,7 +66,7 @@ void solve() {
     int N;
     cin >> N;
 
-    int d, k, s, h;
+    ll d, k, s, h;
     d = k = s = h = 0;
 
     while (N--) {
